sql: disallow id duplicate in insert & update
Before this patch one was able to exec query like this:
  `insert into t(col1, col1) values(1, 1)`
  `update t set col1 = 1, col1 = 2;`
which is nonsense.

New policy is to throw an error in case of duplicates in columns
description.

The fix for insert compares each identifier with the others and therefore
has O(n^2) complexity. However, it seems that to maintain faster data
structure just for this small check is redundant.

The fix for update takes O(n), because it is implemented on initialized
array.

Closes #2358